Fix more type inconsistencies

+++ generator/data.hpp
@@ -66,7 +66,7 @@ inline GLuint textures[TEX_COUNT];
 inline std::string textureOutputDir=".";
 inline GLint transmittanceTexW, transmittanceTexH;
 inline GLint irradianceTexW, irradianceTexH;
-inline glm::vec4 scatteringTextureSize;
+inline glm::ivec4 scatteringTextureSize;
 inline auto scatTexWidth()  { return GLsizei(scatteringTextureSize[0]); }
 inline auto scatTexHeight() { return GLsizei(scatteringTextureSize[1]*scatteringTextureSize[2]); }
 inline auto scatTexDepth()  { return GLsizei(scatteringTextureSize[3]); }
